requirements.txt
redis
aiogram>=3.0.0
gtts>=2.3.2
python-dotenv>=1.0.0
colorama>=0.4.4
